svg/painting/parsing/text-rendering-invalid.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml" width="800" height="800"><title>SVG Painting: parsing text-rendering with invalid values</title><metadata><h:link href="https://svgwg.org/svg2-draft/painting.html#TextRenderingProperty" rel="help"/><h:meta content="text-rendering supports only the grammar 'auto | optimizeSpeed | optimizeLegibility | geometricPrecision'." name="assert"/></metadata><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><h:script src="/css/support/parsing-testcommon.js"/><script><![CDATA[

test_invalid_value("text-rendering", "crispedges");
test_invalid_value("text-rendering", "auto optimizespeed");

  ]]></script></svg>
